+++ Assets/Scripts/Components/GrayOut.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using Services.Implementations;
+using UnityEngine;
 
 public class GrayOut : MonoBehaviour
 {
