+++ Assets/Scripts/Components/Player.cs
@@ -1,6 +1,8 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using Services;
+using Services.Implementations;
 using UnityEngine;
 
 public class Player : EventHandlingComponent
