--- Assets/Scripts/IInputService.cs
@@ -1,20 +1,0 @@
-using UnityEngine;
-
-public interface IInputService
-{
-    float GetAxisRaw(string axisName);
-    bool GetButton(string buttonName);
-}
-
-class UnityInputService : IInputService
-{
-    public float GetAxisRaw(string axisName)
-    {
-        return Input.GetAxisRaw(axisName);
-    }
-
-    public bool GetButton(string buttonName)
-    {
-        return Input.GetButton(buttonName);
-    }
-}
