--- Assets/Scripts/ITimeService.cs
@@ -1,14 +1,0 @@
-using UnityEngine;
-
-public interface ITimeService
-{
-    float GetDeltaTime();
-}
-
-class UnityTimeService : ITimeService
-{
-    public float GetDeltaTime()
-    {
-        return Time.deltaTime;
-    }
-}
