+++ Assets/Scripts/Services.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 98fda86a4541d444aa4b6ce768b6e615
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
