+++ Assets/Scripts/Services/IInputService.cs
@@ -1,0 +1,8 @@
+namespace Services
+{
+    public interface IInputService
+    {
+        float GetAxisRaw(string axisName);
+        bool GetButton(string buttonName);
+    }
+}
