+++ Assets/Scripts/Services/ITimeService.cs
@@ -1,0 +1,7 @@
+namespace Services
+{
+    public interface ITimeService
+    {
+        float GetDeltaTime();
+    }
+}
