+++ Assets/Scripts/Services/Implementations/UnityTimeService.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+namespace Services.Implementations
+{
+    class UnityTimeService : ITimeService
+    {
+        public float GetDeltaTime()
+        {
+            return Time.deltaTime;
+        }
+    }
+}
